+++ src/Generator/Internal.php
@@ -2,8 +2,10 @@
 
 namespace Riimu\Kit\SecureRandom\Generator;
 
+use Riimu\Kit\SecureRandom\GeneratorException;
+
 /**
- * Generates bytes using PHP's built in CSPRNG.
+ * Generates bytes and numbers using PHP's built in CSPRNG.
  *
  * PHP7 offers a built in function for generating cryptographically secure
  * random bytes. This class simply wraps that method for supported PHP versions.
@@ -12,7 +14,7 @@ namespace Riimu\Kit\SecureRandom\Generator;
  * @copyright Copyright (c) 2014, Riikka Kalliomäki
  * @license http://opensource.org/licenses/mit-license.php MIT License
  */
-class Internal extends AbstractGenerator
+class Internal extends AbstractGenerator implements NumberGenerator
 {
     public function isSupported()
     {
@@ -22,5 +24,24 @@ class Internal extends AbstractGenerator
     protected function readBytes($count)
     {
         return random_bytes($count);
+    }
+
+    public function getNumber($min, $max)
+    {
+        $min = (int) $min;
+        $max = (int) $max;
+        $exception = null;
+
+        try {
+            $number = random_int($min, $max);
+        } catch (\Throwable $exception) {
+            $number = false;
+        }
+
+        if (!is_int($number) || $number < $min || $number > $max) {
+            throw new GeneratorException('Error generating random number', 0, $exception);
+        }
+
+        return $number;
     }
 }