+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changelog #
 
+## v1.2.0 (2016-05-11) ##
+
+  * Added support for PHP's internal CSPRNG in php 7.0 (which is used by default)
+  * Added NumberGenerator interface for generators that can natively generate numbers
+
 ## v1.1.2 (2015-08-14) ##
 
   * Address some unlikely corner cases